Adds basic node structure
Introduces three classes: Node, RootNode, LeafNode.
The first is a generic node used in the majority of the tree
decomposition. The second contains supplementary information regarding
the simulation, can run iterative steps of the whole tree, perform
decomposition operations etc...
The last is located at the bottom of the tree and contains informations
relevant to the Particle's located in this domain. It can also perform
direct summation operations within itself or its direct neighbours.