Adds Partial Decomposition Mechanisms
A method now handles setting the domains boundaries. When creating a
(sub)node, each parent is responsible for setting its children's
boundaries. The RootNode handles setting its boundaries within its
constructor.
Instead of using dynamically arrays to store particles, we switch to
vectors of pointers to dynamically allocated particles, so that the
decomposition operations are achieved easily. The input mechanisms are
consequentially updated. The process of freeing the memory is still
performed in the SConfig destructor (also the one to allocate them).
A basic error management system is layed out where different errors are
categorized within an enumeration.
The particles now have an additional field to store their force. It is
initialized to zero at the beginning of the simulation (to prevent the
user from overriding the FMM's computation).