Adds a Routine Browsing Isolevel Nodes
The function Node::GetNext() returns a pointer to a Node object on the
same level of refinement. Following the Nodes with the current
decomposition algorithm yields a so-called Z-curve in two dimensional
space. The routine answers NULL if no Node can be found after the
calling one.

The function RootNode::Run() contains a small snippet demonstrating the
routine's behavior.