Adds a Movement Function for 2D Displacements
In order to perform nearest neighbours finding tasks, a new function is
required. The tree decomposition allows stepping in and out of levels
easily by following children or parents, but back in the 2D plane, this
decomposition can also make two physically close nodes very distant in
the memory:
For example, take a LeafNode with a corner at the center of the domain.
The Leafs sharing this same corner were separated when creating the
first level below the RootNode.

The Node::Move method takes a direction, looks for the node in this
direction relative to the caller and returns a pointer to it. If no Node
can be found (either because we're at the Root, or against a wall), the
method returns a NULL pointer.