Adds Time Evolution and Particle Reassignements
The time evolution method is implemented using the Euler scheme for the
velocity and position of the particles. To preserve tree locality, each
particle is evolved directly by its leaf node.

After evolving a particle, the time evolution method checks whether a
domain reassignement is required. If so, the particle is deleted from
the local leaf and propagated upwards in the tree, until it a node from
which it can be propagated downwards again is found.

This method is called for each particle and will likely represent a
bottleneck for future versions. A smarter approach would be to perform a
verification pass after the tme evolution and dispatch the particles to
new domain by groups, or to dispatch them from the root no matter what.

It was found that using floating point representation for the domain
edges could lead to imprevisible behavior where a particle becomes stuck
between domain walls. To avoid such an issue, the positions of the
simulated objects have been rolled back from float to double.

Running large simulations, a stack related issue appeared once, but its
origin wasn't found, so keep an eye on this. A flag was added to the
Makefile for this purpose.