MATLAB Helpers Update
Adds support for varying particle masses, more initial profiles and an
error computation routine, force_error.
force_error computes the expected force acting on a requested particle
at each timestep (depending on the distribution of the rest of the
masses) and returns the difference between the simulated force and the
expected one.