Fixes in Time Evolution
The previous code was computing the forces acting on each particle and
updating their positions and velocities right away. This is of course
wrong, since we manipulate only a single copy of these objects, and
those properties must be updated globally at once to prevent numerical
errors. A secondary function is implemented (TimeEvolutionMasses) to
perform the updating task.