Implement Cosntructors for Nodes
Rework initialization of Nodes. A root node (equivalent of the previous
RootNode class) can be called with a public constructor, which requires
the SConfig input from the user. Another constructor is used internally
by the class to add children to the tree. It will be called by the Nodes
when initializing and updating the tree and automatically sets
geometrical boundaries, level and index of the node according to its
parent.

Also add two file which will contain the routines of the simulation
(initialization, tree building, time evolution) that are meant to be
called internally or from the main.