Adds Higher Level Data Structures
Removes some methods from the Node class, and adds the QuadTree class.
This class will act as a wrapper for Node objects, keep a reference to
the root of the tree, and provide methods to modify it (add, remove or
move particles around).

The simulation class is introduced and will contain the routines needed
to perform a whole run, including IO-related calls.