Fixes to BH Tree Decomposition
The tree decomposition occurs without crash on a small particle system
(10 elements). Multiple fixes are provided:

- The Simulation class only takes a reference to the SConfig object to
avoid copies of the dynamically initialized array of particles. Else, a
double free occurs.
- Dynamically allocated elements are freed, and their pointer is set to
NULL (or deleted in memory) to avoid double frees.
- The private Node constructor asked the parent class which child it
was, but this assumed the assignement of the pointer in the parent class
was already performed, which is not the case in most applications. The
constructor now asks for its quadrant (child) index and works with this.