Adds Fictious Nodes to the Tree
In earlier versions, pointers to particles were being passed down the
tree on insertion, but when the tree needed to be deepened at a certain
node, the pointer wasn't replaced.
This is naturally problematic since a change at a given level (saya the
leaf), would propagate in all the tree.

Fictious nodes are created with a "-1" identifier so not to mix them
with the simulated ones. The next commits will focus on tree updates.