Adds Leaf Browsing Method
This method is based on the tree browser implemented in the QuadTree
operator<< overloaded function. It returns the next leaf it finds in the
tree. If no leaf is found, or if a NULL pointer is passed, a NULL
pointer is returned. If the argument pointer is not a leaf, the browser
moves to the leaf level automatically. This allows starting the routine
easily.