Introduce Force Computation
In the Simulation class, a force computation routine is introduced.
All leafs in the tree are browsed using the available method in the
QuadTree class. For all leafs, the routine iterates over tree nodes, and
recursively explores them depending on the theta parameter of the
Barnes-Hut model (accuracy of the force). Previously computed forces on
leafs are reset by this routine.

Once a Node is reached with a precision deemed sufficient for force
computation, the interaction between the leaf node (that is, the
particle) and the center of mass of the current node is computed.

The force computation routine involves methods mostly independent of the
Node classes and Particle structures. Mechanisms avoiding
self-interaction are provided. The input file scanner is modified to
parse the theta parameter.

It is verified the forces are finite (meaning self-interaction is
avoided), but the overall accuracy of the scheme remains to be verified
(a consistent IO routine would help).