First Functional IO
Prints LeafNodes using the default constructor.
Warning: Has very (very) dirty std::ofstream  manipulation routines,
which pretty much assume input and output files are to be read in a
separate folder.

The main also does not check the number of arguments available before
calling the IOManager constructor.
The class IOManager should have easy access to the data structures. We
need to overload an operator and possibly friend it.