Iteration-Wise Output for Particles
Outputs particle informations in the CSV format. To speed up the output
process, the particles are not ordered (we pick whatever order the tree
imposes, and do not sort the particles ID, which are print).

The routine is checked not to output fictious node's data (-1 indexes).
The IOManager class is friended to the Node class to access the center
of mass data easily.