Adds an Euler Scheme for Time Evolution
This method iterates over leaf nodes starting from the root of the
quad-tree stored in the simulation object. We choose to perform the
iteration this way to avoid depending on the particles stored in the
SConfig object (which could be difficult to keep in a parallel
implementation). A public method Node::GetParticle() returns a pointer
to the particle representation of a node (will potentially simplify
future calls, such that friend dependencies can be removed later).