Update to GetNext Method
This method was imported from earlier versions of the code (<0.6), and
was designed for an uniform tree. Thus, the method would for example
only check that the current's node index (within its parent) was not
three before returning the child on the same level.

We fix the method by adding checks for NULL pointers and enable
"partial" tree scrolling.