Adds Particle Reassignement Routine
Begins completing the Simulation::UpdateTree method. The aimed-for
structure of this routine is a two-step tree update mechanism,

1. All leafs are browsed, and Particles that escaped their host node's
coverage area are reassigned.
2. All empty Nodes are detected, and pruned.

This commit implements the first element above. A
QuadTree::RemoveParticle method is first called to remove the particle
from the tree (which is updated until the root, to account for center of
mass changes). Then, the AddParticle method used for initial tree
building is called on the same particle.

The current implementation expects non-fictive particles to be passed,
i.e. with index different from -1, but does not check it rigorously.