Adds Node Pruning Mechanisms
After reassigning particles to matching domains upon a time step, a
mechanisms is implemented to manage nodes that were emptied. We call it
a pruning mechanism: it removes unnecessary nodes (empty, deletes them)
and eventually pulls remaining children nodes up the tree to more
shallow levels.

A caveat of this mechanism is that pruning a node can radically change
the structure of the local tree, such that, upon pruning a node, the
whole tree is rescanned (in the current implementation at least, more
efficient mechanisms may be proposed later on).