Quick Fix when only Two Particles are Simulated
Adds small checks to avoid a segmentation fault when two particles
remain in the system, and one is exiting it. Although a segmentation
fault occurs the following time step, we have a last clean call to IO
routines before the SIGSEV and may allow us to cleanly exit the main
loop when error management is added in.