Implement a Resized MPI Datatype for Particles
When using the MPI Datatype for Particle as introduced in the previous
commit, we are only allowed to transfer one particle at the time. This
is due to the way the compiler will align memory for Particle objects:

* A double is 8-bytes aligned. We use it for representing the position,
velocity, force and mass (7x) of a particle.
* A (signed) long is 4-bytes aligned. We use it for the particle index.

Thus, it could be that the compiler allocates 4 "hollow" bytes at the
end of the Particle structure so that its total size matches 4*8=32
bytes (the size of a word). We indicate the MPI library that this type
of allocation may happen by "resizing" the type when the MPI type size
differs from the original one.