Increase the scope of MPI_Particle
Since the MPI type for a particle will be needed in multiple routines
through the code, the type is defined globally in stypes.hpp, and
`extern` mentions of it are added in the necessary files.

The code to set up and free the type is moved to a separate function
called from the main.