Adds Configuration Broadcasting
The IOManager class now has a configuration broadcasting method
available. Contrarily to the case of Particles, we do not create an MPI
type corresponding to a SConfig object, since our configuration's
largest attributes are the most heavy and useless for threads other than
the master.
Thus, a buffer is allocated, and the configuration parameters of type
double and unsigned are sent through a regular broadcast instruction.

Remark: The aforementioned "largest attributes" consist of the vector of
Particle pointers, as well as the strings containing input and output
paths. We want to pass particles selectively, and we only plan to have
the master thread writing (rank=0), so these parameters are deemed
useless and left uninitialized in the processes with rank>0.