Fix MPI_Particle and Finalize Tree Distribution
Fixes the number of fields in MPI_Type_create_struct to map all needed
properties of the class (only positions and velocitities were sent).

The tree distribution routine is moved to the IOManager class, so that
the configuration reference in the Simulation object is kept constant.
It also allows to clearly separate the communication aspects of the
process.

The tree distribution routine now allocates particles received through
the MPI buffer on the heap (call to new) for each child thread". Another
block, executed after the send instruction on CPU0, clears the memory
associated with these particles (it doesn't clear the tree yet).

Debugging messages are kept in the routine for later on.
We now focus on partial (thread-wise) tree building and Particle
gathering to the root process for occasional full tree rebuilds.