Clear Pointer in Nodes which COM is Removed
When distributing the tree, particles that are sent to other processors
are removed from the local process (deleted, since allocated on the
heap). We remove the reference to this center of mass in affected nodes
(a tree was already built!) since the Node destructor tries to perform
this operation too.