Enhance Tree Update Mechanism
Up to this version, every time a Node was pruned in the
Simulation::TreeUpdate function, the "update pointer" was reset to the
first leaf found from the root of the tree. This mechanism performs o.k.
with small tree size, but quickly became unbearable for large ones (10^5
particles and more).

Instead of restarting from the root, we query a restart node from the
QuadTree::PruneNode function, which may have been moved by some levels
up the tree, depending on the importance of the pruning. UpdateTree
moves the received pointer to the leaf level and then continues.