Add Shortcut in Leaf Syncing Routine
The IOManager::SyncLeafs routine sends particles managed by the running
threads to other processes executed in parallel. It also receives
particles from these same processes.

When a particle is received, the routines queries its ID and browses the
list of particles stored in the configuration to find a matching
particle, and update it.

For all cases so far, each thread stores a similar configuration, with
particles ordered with increasing index, such that we deem possible to
override the index comparaison and access the concerned particle
directly.

This fix will fail if:

* A way to assign the index is provided in the input files (such that
the user is responsible for this choice, and not the program itself).
* Each thread stores a different set of particles.