Add Timeout to Simulation::Run
Uses the MPI_Wtime to detect whether a run is exceeding its allocated
time. Will be used on clusters when a timeout can be passed to the SLURM
scheduler.

Currently, the exit criterion is verified thread-wise, which could be
problematic if the clocks are not synchronised. Will fix later if
necessary.