Initial Performance Fix Commit
Now that I finally understand how to run VTune alongside with the SLURM
tool, I am able to see some bottlenecks in the code. This branch is
aimed at small performance fixes (requiring only a few commits). If
larger fixes are needed, a separate branch will be created (do not
expect it however, since fixing the code is not my main concern right
now).

In this commit, we avoid copying the center of mass of a particle in the
associated Node::GetCenterOfMass() function.