Moves to Intel Compilers and Includes VTune Analysis
Instead of loading gcc and mvapich2, we now use the intel and intel-mpi modules
from the cluster. We fix the include preprocessors instructions in the STypes
source file (to use offsetof), change the name of the compiler in the Makefile,
and set debug flags back on (because we're optimising! yey).